refactor(mandate-step2): extract mandate_store tests mechanically
